[FIX] QuickActions for mobile screen (#23016)
* WIP

* Fix WIP

* Fix

* lint

* use filtered action

* fix actions in mobile
