+++ client/views/room/Header/ToolBox/ToolBox.tsx
@@ -1,6 +1,6 @@
 import { Menu, Option, MenuProps, Box } from '@rocket.chat/fuselage';
 import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
-import React, { memo, useContext, ReactNode, useRef, ComponentProps, FC } from 'react';
+import React, { memo, ReactNode, useRef, ComponentProps, FC } from 'react';
 // import tinykeys from 'tinykeys';
 
 // used to open the menu option by keyboard
@@ -9,7 +9,7 @@ import Header from '../../../../components/Header';
 import { useLayout } from '../../../../contexts/LayoutContext';
 import { useTranslation } from '../../../../contexts/TranslationContext';
 import { ToolboxActionConfig, OptionRenderer } from '../../lib/Toolbox';
-import { ToolboxContext } from '../../lib/Toolbox/ToolboxContext';
+import { useToolboxContext } from '../../lib/Toolbox/ToolboxContext';
 import { useTab, useTabBarOpen } from '../../providers/ToolboxProvider';
 
 const renderMenuOption: OptionRenderer = ({ label: { title, icon }, ...props }: any): ReactNode => (
@@ -28,7 +28,8 @@ const ToolBox: FC<ToolBoxProps> = ({ className }) => {
 	const t = useTranslation();
 	const hiddenActionRenderers = useRef<{ [key: string]: OptionRenderer }>({});
 
-	const { actions: mapActions } = useContext(ToolboxContext);
+	const { actions: mapActions } = useToolboxContext();
+
 	const actions = (Array.from(mapActions.values()) as ToolboxActionConfig[]).sort(
 		(a, b) => (a.order || 0) - (b.order || 0),
 	);
