+++ client/views/room/lib/QuickActions/QuickActionsContext.tsx
@@ -1,17 +1,20 @@
 import { EventHandlerOf } from '@rocket.chat/emitter';
-import { createContext } from 'react';
+import { createContext, useContext } from 'react';
 
 import { actions, listen, QuickActionsActionConfig, QuickActionsAction, Events } from '.';
+
 import './defaultActions';
 
 export type QuickActionsEventHandler = (handler: EventHandlerOf<Events, 'change'>) => Function;
 
-export type ChannelContextValue = {
+export type QuickActionsContext = {
 	actions: Map<QuickActionsActionConfig['id'], QuickActionsAction>;
 	listen: QuickActionsEventHandler;
 };
 
-export const QuickActionsContext = createContext<ChannelContextValue>({
+export const QuickActionsContext = createContext<QuickActionsContext>({
 	actions,
 	listen,
 });
+
+export const useQuickActionsContext = (): QuickActionsContext => useContext(QuickActionsContext);
