+++ client/views/room/lib/Toolbox/ToolboxContext.tsx
@@ -1,12 +1,12 @@
 import { EventHandlerOf } from '@rocket.chat/emitter';
-import { createContext, MouseEventHandler } from 'react';
+import { createContext, MouseEventHandler, useContext } from 'react';
 
 import { actions, listen, ToolboxActionConfig, ToolboxAction, Events } from '.';
 import './defaultActions';
 
 export type ToolboxEventHandler = (handler: EventHandlerOf<Events, 'change'>) => Function;
 
-export type ChannelContextValue = {
+export type ToolboxContextValue = {
 	actions: Map<ToolboxActionConfig['id'], ToolboxAction>;
 	listen: ToolboxEventHandler;
 	tabBar?: any;
@@ -17,10 +17,12 @@ export type ChannelContextValue = {
 	activeTabBar?: ToolboxActionConfig;
 };
 
-export const ToolboxContext = createContext<ChannelContextValue>({
+export const ToolboxContext = createContext<ToolboxContextValue>({
 	actions,
 	listen,
 	open: () => null,
 	openUserInfo: () => null,
 	close: () => null,
 });
+
+export const useToolboxContext = (): ToolboxContextValue => useContext(ToolboxContext);
